Refactor Vulkan sync and image handling, add helpers
Replaces VulkanInit.h with VulkanInitInfo.h/cpp, introducing helper functions for Vulkan object creation and synchronization. Adds VulkanImage.h/cpp for image layout transitions using vkCmdPipelineBarrier2 and updates Renderer to use new synchronization primitives and image transitions. Updates VulkanDevice to enable VK_KHR_synchronization2 and related features/extensions, and modifies swapchain image usage flags. Cleans up and modernizes device creation and frame synchronization logic.

+++ Quasar/src/Renderer/Renderer.h
@@ -26,7 +26,7 @@ namespace Quasar
         FrameData& get_current_frame() { return _frames[_frame_number % FRAME_OVERLAP]; };
 
         private:
-        u8 _frame_number {0};
+        u32 _frame_number {0};
         FrameData _frames[FRAME_OVERLAP];
 
         u32 _api_major; // The instance-level api major version.
